clk: tegra: Add function to get the CPU rail alignment
The CPU rail alignment is used for rounding the CPU rail voltages.
Currently, the DVFS driver uses static pre-defined values for the
rail alignment and not those from device-tree that are read by the
DFLL driver. Therefore, update the DFLL driver to store all the
alignment data in the SoC structure and add a function to allow the
DVFS driver to retrieve the alignment data from the DFLL driver.

This corrects some rounding errors for the CPU voltage rail when
computing the freq/voltage table.

Bug 1811732

Change-Id: Ie9f38cc321c04a63598743dde3d0d249e25a7170
Signed-off-by: Jon Hunter <jonathanh@nvidia.com>
Reviewed-on: http://git-master/r/1284919