v4l2: Increase event subscription queue size
Increase the size of event subscription queue to 2.

If the internal event queue for a particular subscribed event
becomes full, then the oldest event in that queue will be dropped.
Since we need both the on and off signals to our subscription,
it is good to increase the queue to 2.

Bug 200602400

Change-Id: Ic38517639448b10e7723d18ae91e1c6e793a0563
Signed-off-by: Gigon Bae <gbae@nvidia.com>
(cherry picked from 677e58eb417af5ac4d72e630aa52acd5647a0933)
Reviewed-on: http://git-master/r/1294535
Reviewed-on: http://git-master/r/1469821
Reviewed-on: https://git-master.nvidia.com/r/1662440
Reviewed-on: https://git-master.nvidia.com/r/1740237
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2401277
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Bitan Biswas <bbiswas@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: Ashwini Purkar <apurkar@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>