dc: Disable all active windows on suspend
Commit e746350ed9c2 ("tegra: video: blank all valid windows")
added logic to tegra_dc_disable_irq_ops() so that it disables
all active windows for a given head before engaging in the SOR
detach sequence, regardless of the dc_ext ownership status of
those windows.

The same fix needs to be applied to the tegra_dc_suspend()
path as well.

Bug 200485788

Change-Id: I371662c772ed6fdd8032afdbfb43c7864ae0cb79
Signed-off-by: Shu Zhong <shuz@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/2103894
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Naveen Kumar S <nkumars@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Ujwal Patel <ujwalp@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>