cpufreq: create a single policy
create a single policy for all 8 CPUs, so that the frequency
change is in lock-step for all CPUs.

This is controlled by property for device tree node "cpufreq":
cpufreq_single_policy. When this property is present, a single
policy is created. If not, one policy is created for each cluster.

The change helps to improve scaling in case of workloads that
have high inter-dependency when running on different clusters.

Bug 200674300

Change-Id: Idfba13acfdb6c73b368e463d731f008ffc7abd49
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1710748
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2578478
Reviewed-by: svcacv <svcacv@nvidia.com>
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sumit Gupta <sumitg@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit