reset: take reset_control_deassert in reverse order
The reset signal propagation usually needs to take care from the bus or
memory controller to the peripheral clients. When asserting the reset
signal, this needs to start with bus or memory controller to clear the
pending or onging IO process, then reset the peripheral.

And needs to take care the reset deassertion in reverse order to avoid
the IO accessing on the bus or memory but it still in reset state. So
this patch provides the solution when we need to take care of the reset
signal propagation in the topology we mentioned.

Bug 200416069

Change-Id: I686df004abf325746756db32488d66ec9a049f1d
Signed-off-by: Joseph Lo <josephl@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1747055
(cherry picked from commit 85cfc5e75921f7706db96062023091845f0f0580)
Signed-off-by: Nicolin Chen <nicolinc@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2444434
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit