kernfs: sanity check pointer
The strlcpy function returns the length of source pointer when
the requested size is 0. this behavior is relied upon for sched
tracing. We can't simply return when buf is 0, but we have to
protect against the scenario when buf is 0 and requested size
is non-0, in which case the strlcpy would lead to illegal memory
access.

bug 200422707
bug 200426751

Change-Id: I0cad6dd53d1fa530a8d339d4621954a4833d2ad9
Signed-off-by: Bo Yan <byan@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1776362
(cherry picked from commit 22c11173cdc3b9386bf8049aa8ded9f6c32cf432)
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.9/+/2411255
Reviewed-by: automaticguardword <automaticguardword@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>