fbcon: Warn if no matching mode can be set
var_to_display() should always be able to find a matching node because
the content of struct fb_var_screeninfo was derived from a mode in the
modelist associated with the struct fb_info.

However there are cases where this can happen: when a mode is converted
to a struct fb_var_screeninfo and back again, rounding errors can result
in the mode no longer matching the original. When this happens it causes
a crash further down the road in fbcon_switch() because it will try to
access a NULL mode. To make matters worse, this crash happens with the
console lock held, making it almost impossible to diagnose without going
in and disabling the lock.

Add a warning if this case is detected to give a hint about why things
might be going wrong.

Bug 200130894

Change-Id: I27bc0dca8d7b4d4d4dafb3d8829a5940a2f28995
Signed-off-by: Thierry Reding <treding@nvidia.com>
Signed-off-by: Naveen Kumar S <nkumars@nvidia.com>
Reviewed-on: http://git-master/r/929901
(cherry picked from commit 284352fe92266299af6281e32d84ca25d8d2fb1d)
Reviewed-on: https://git-master.nvidia.com/r/1787463
GVS: Gerrit_Virtual_Submit
Reviewed-by: Sachin Nikam <snikam@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>