sched: cpufreq: sysfs knob for capacity_margin
This commit makes the capacity_margin knob configurable from user-space.
By tinkering with this knob, cpufreq scaling of frequency can be made
more aggressive. Allowed values for this knob range from 0 to 1024. By
default, it has the value of 256 which sets the frequency scaling factor
to be 1.25

Bug 200674300

Change-Id: I76138a3668f28548e6d08ae5321c1b1d9ce6f39d
Signed-off-by: Waqar Ali <waqara@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1794474
Reviewed-on: https://git-master.nvidia.com/r/c/linux-5.10/+/2578481
Reviewed-by: svc_kernel_abi <svc_kernel_abi@nvidia.com>
Reviewed-by: Bibek Basu <bbasu@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: Sumit Gupta <sumitg@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>
GVS: Gerrit_Virtual_Submit