video: tegra: host: fix prepare_suspend
The usage_count could be greater than 1 if an asynchronous suspend
was triggered while the device is in runtime suspend i.e. prepare
count is one. Remove the usage_count check and allow device to
suspend if it doesn't have any channels in use.

Avoid disabling the clocks in runtime_suspend if the device was idle
previously.

Bug 200358755
JIRA HOSTX-335
Bug 2356911

Change-Id: I0b83b57f8c6149d34096e5089e3bfa4cbba6e922
Signed-off-by: Srikar Srimath Tirumala <srikars@nvidia.com>
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1817554
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Damian Halas <dhalas@nvidia.com>
Tested-by: Damian Halas <dhalas@nvidia.com>
Reviewed-by: Nirav Patel <nipatel@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>