video: tegra: host: Give jobs some time to finish when suspending
Currently if there are any jobs in the sync queue while the system is
entering system suspend, the suspension is cancelled. However, we
should be able to without problems wait a little bit for anything
ongoing to finish.

Do just that by waiting at most 10 times for a 1ms period for all jobs
to finish. If they still aren't at that point, cancel the suspension.

Bug 2356911
Jira HOSTX-373

Change-Id: I950e4a00a56281fb0b46fa45c9c6030755aba183
Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1821872
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Damian Halas <dhalas@nvidia.com>
Tested-by: Damian Halas <dhalas@nvidia.com>
Reviewed-by: Nirav Patel <nipatel@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>