ASoC: tegra-alt: Boot the ADSP OS during probe
Currently the ADSP OS is booted only when the mixer control
'ADSP Init' is set, So ADSP OS won't be booted and during
alsa-restoring of service it lead to os not booted error
message.

So try booting the ADSP OS as soon as ADSP tegra-alt driver
is registered.

Bug 200458098

Change-Id: I3cd123bdb5d5f87ba8c9329d0e43241189ffdb66
Signed-off-by: Mohan Kumar <mkumard@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1935089
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: Jonathan Hunter <jonathanh@nvidia.com>
Tested-by: Jonathan Hunter <jonathanh@nvidia.com>
Reviewed-by: svc-mobile-coverity <svc-mobile-coverity@nvidia.com>
Reviewed-by: svc-mobile-misra <svc-mobile-misra@nvidia.com>
GVS: Gerrit_Virtual_Submit
Reviewed-by: Hariharan Sivaraman <hariharans@nvidia.com>
Reviewed-by: Viswanath L <viswanathl@nvidia.com>
Reviewed-by: Nitin Pai <npai@nvidia.com>
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>