dts: t19x: move bpmpthermal node to soc-base
This change is preparing the introduction of upstream compatible device
tree bindings for bpmp-fw. For upstream bindings, it's needed that the
bpmp node is declared after it's dependencies for correct probing.

Hence, move all bpmp subdevices to tegra194-soc-base.dtsi file to
avoid bpmp to be introduced too early in the device tree blob.

Bug 2416725

Change-Id: I7fee24fb18c1fb376139352138d1024219d66db7
Signed-off-by: Timo Alho <talho@nvidia.com>
Reviewed-on: https://git-master.nvidia.com/r/1943311
Reviewed-by: Automatic_Commit_Validation_User
Reviewed-by: mobile promotions <svcmobile_promotions@nvidia.com>
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>